#35: Split tf-data-code into more packages
