+++ dsl/src/main/kotlin/edu/wpi/axon/dsl/KoinModules.kt
@@ -6,12 +6,12 @@ import edu.wpi.axon.dsl.validator.path.DefaultPathValidator
 import edu.wpi.axon.dsl.validator.path.PathValidator
 import edu.wpi.axon.dsl.validator.variablename.PythonVariableNameValidator
 import edu.wpi.axon.dsl.validator.variablename.VariableNameValidator
-import edu.wpi.axon.tfdata.code.DefaultLayerToCode
-import edu.wpi.axon.tfdata.code.DefaultLossToCode
-import edu.wpi.axon.tfdata.code.DefaultOptimizerToCode
-import edu.wpi.axon.tfdata.code.LayerToCode
-import edu.wpi.axon.tfdata.code.LossToCode
-import edu.wpi.axon.tfdata.code.OptimizerToCode
+import edu.wpi.axon.tfdata.code.layer.DefaultLayerToCode
+import edu.wpi.axon.tfdata.code.loss.DefaultLossToCode
+import edu.wpi.axon.tfdata.code.optimizer.DefaultOptimizerToCode
+import edu.wpi.axon.tfdata.code.layer.LayerToCode
+import edu.wpi.axon.tfdata.code.loss.LossToCode
+import edu.wpi.axon.tfdata.code.optimizer.OptimizerToCode
 import org.koin.dsl.module
 
 fun defaultModule() = module {
