+++ tf-data-code/src/main/kotlin/edu/wpi/axon/tfdata/code/optimizer/OptimizerToCode.kt
@@ -1,4 +1,4 @@
-package edu.wpi.axon.tfdata.code
+package edu.wpi.axon.tfdata.code.optimizer
 
 import edu.wpi.axon.tfdata.optimizer.Optimizer
