+++ tf-data-code/src/test/kotlin/edu/wpi/axon/tfdata/code/layer/DefaultLayerToCodeTest.kt
@@ -1,5 +1,6 @@
-package edu.wpi.axon.tfdata.code
+package edu.wpi.axon.tfdata.code.layer
 
+import edu.wpi.axon.tfdata.code.layer.DefaultLayerToCode
 import edu.wpi.axon.tfdata.layer.Activation
 import edu.wpi.axon.tfdata.layer.Layer
 import edu.wpi.axon.tfdata.layer.SealedLayer
