+++ tf-data-code/src/test/kotlin/edu/wpi/axon/tfdata/code/optimizer/DefaultOptimizerToCodeTest.kt
@@ -1,5 +1,6 @@
-package edu.wpi.axon.tfdata.code
+package edu.wpi.axon.tfdata.code.optimizer
 
+import edu.wpi.axon.tfdata.code.optimizer.DefaultOptimizerToCode
 import edu.wpi.axon.tfdata.optimizer.Optimizer
 import io.kotlintest.shouldBe
 import org.junit.jupiter.params.ParameterizedTest
